requirements.txt
requests
fake-useragent
chardet
chardet
pyyaml
